move find_parcels_within_complex function as a method to the ParcelsWithinComplexFinder class
